Fixing SBSA MP usage (#4)
The existing TFA will fault the secondary cores when placing the TFA image in secure flash. This was because when the stmm image is trying to communicate with the secure flash for UEFI variable support, the flash will change mode to write. In the meantime, the secondary cores looping at `wfe` will be woken up which at this point the flash cannot be read.

This change will need to pair with PROGRAMMABLE_RESET_ADDRESS=1 build option to hold the secondary cores to loop at BL31 instead of BL1 which will remove the execute on flash dependency to avoid such issue.

Note that the holding pen logic is also updated to specifically handle the case of context losing suspend, where a wfi is used which makes the core not to wake up more until an explicit interrupt is issued.